requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.13.0
APScheduler==3.11.2
asyncio-throttle==1.0.2
attrs==26.1.0
audioread==3.1.0
babel==2.18.0
bangla==0.0.5
bcrypt==5.0.0
beautifulsoup4==4.14.3
bitsandbytes==0.49.2
blinker==1.9.0
blis==1.3.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
build==1.4.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.7
click==8.3.2
cloudpathlib==0.23.0
confection==1.3.3
contourpy==1.3.3
coqpit==0.0.17
cryptography==46.0.7
cuda-bindings==13.2.0
cuda-pathfinder==1.5.3
cuda-toolkit==13.0.2
cycler==0.12.1
cymem==2.0.13
Cython==3.2.4
datasets==4.8.4
dateparser==1.1.8
decorator==5.2.1
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
docopt==0.6.2
durationpy==0.10
einops==0.8.2
encodec==0.1.1
fastapi==0.135.3
filelock==3.28.0
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
g2pkk==0.1.2
googleapis-common-protos==1.74.0
GPUtil==1.4.0
greenlet==3.4.0
grpcio==1.80.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.10.2
idna==3.11
impacket==0.13.0
importlib_metadata==8.7.1
importlib_resources==7.1.0
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonlines==1.2.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kubernetes==35.0.0
lazy-loader==0.5
librosa==0.11.0
linkify-it-py==2.1.0
llama_cpp_python==0.3.20
llvmlite==0.47.0
loguru==0.7.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdit-py-plugins==0.5.0
mdurl==0.1.2
mmh3==5.2.1
more-itertools==11.0.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.19
murmurhash==1.0.15
networkx==2.8.8
nltk==3.9.4
num2words==0.5.14
numba==0.65.0
numpy==1.26.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
oauthlib==3.3.1
onnxruntime==1.24.4
openai==2.31.0
openai-whisper==20250625
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.8
overrides==7.7.0
packaging==26.1
pandas==1.5.3
paramiko==4.0.0
peft==0.19.0
pillow==12.2.0
platformdirs==4.9.6
playwright==1.58.0
pooch==1.9.0
preshed==3.0.13
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pybase64==1.4.3
pycparser==3.0
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
pyee==13.0.1
Pygments==2.20.0
pynndescent==0.6.0
pyparsing==3.3.2
PyPika==0.51.1
pypinyin==0.55.0
pyproject_hooks==1.2.0
pysbd==0.3.4
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-nmap==0.7.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
rpds-py==0.30.0
safetensors==0.7.0
scapy==2.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
shellingham==1.5.4
six==1.17.0
smart_open==7.6.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
starlette==1.0.0
SudachiDict-core==20260116
SudachiPy==0.6.11
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
textual==8.2.3
thinc==8.3.13
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
torchaudio==2.11.0
tqdm==4.67.3
trainer==0.0.36
transformers==5.5.4
triton==3.6.0
trl==1.1.0
TTS==0.22.0
typeguard==4.5.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uc-micro-py==2.0.0
umap-learn==0.5.12
Unidecode==1.4.0
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
weasel==1.0.0
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.8
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.1
